pyproject.toml
[project]
name = "backend-runai"
version = "0.1.0"
description = "Gemini HTTP proxy for RunAI CLI"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "google-generativeai>=0.8.0",
    "pydantic-settings>=2.1.0",
    "structlog>=24.1.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
